requirements.txt
# gen3==4.9.1
click_loglevel==0.4.0.post1
tabulate==0.8.9
requests==2.27.1
setuptools==62.2.0
# pandas
# numpy
fhirclientr4

pydantic~=1.9.0
pydantic_yaml
pytest==7.1.2

inflection~=0.5.1

networkx==2.8


# dictionaryutils 3.4.4 depends on PyYAML<6.0 and >=5.1
# pypfb 0.5.18 depends on click<8.0.0 and >=7.1.2
# pypfb 0.5.18 depends on importlib_metadata<2.0.0 and >=1.3.0; python_version <= "3.9"
# pypfb 0.5.18 depends on fastavro<=1.4.1

PyYAML==5.4.1
click==7.1.2
importlib-metadata==1.7.0
fastavro<=1.4.1


pypfb==0.5.18
dictionaryutils==3.4.4

matplotlib==3.5.2
flatten_json==0.1.13
